+++ src/main/java/ru/vyarus/dropwizard/guice/injector/InjectorFactory.java
@@ -14,6 +14,11 @@ public interface InjectorFactory {
 
     /**
      * Creates an injector instance.
+     * <p>
+     * NOTE: if overriding modules were used
+     * ({@link ru.vyarus.dropwizard.guice.GuiceBundle.Builder#overrideModules(Module...)}) then modules list will
+     * contain only one overridden module. {@link com.google.inject.util.Modules#override(Module...)} can be used on
+     * this module too, if required.
      *
      * @param stage   target injector stage.
      * @param modules modules supplied to injector