+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 * Update to guice 4.2.0
     - remove guice-multibindings dependency as it's moved to guice core
 * Ability to disable: guicey bundles, guice modules (directly registered) and extensions through bundle builder
-* Add generic disable method to builder: disable(Predicate)     
+* Add generic disable method to builder: disable(Predicate)
+* Add direct support for guice bindings override (using Modules.override() internally): bundle_builder.overrideModules()     
 
 ### 4.1.0 (2017-05-09)
 * Update to dropwizard 1.1.0